# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_TobiasRV_challenge-fs-senior_internals_interfaces.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        type: string
    type: object
  github_com_TobiasRV_challenge-fs-senior_internals_interfaces.LoginResponse:
    properties:
      accessToken:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      refreshToken:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_models.User'
    type: object
  github_com_TobiasRV_challenge-fs-senior_internals_interfaces.LogoutResponse:
    properties:
      message:
        example: Successfully logged out
        type: string
    type: object
...
  github_com_TobiasRV_challenge-fs-senior_internals_interfaces.RefreshTokenRequest:
    properties:
      refreshToken:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  github_com_TobiasRV_challenge-fs-senior_internals_interfaces.RefreshTokenResponse:
    properties:
      accessToken:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
...
  github_com_TobiasRV_challenge-fs-senior_internals_models.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      role:
        $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_models.Userroles'
      teamId:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
  github_com_TobiasRV_challenge-fs-senior_internals_models.Userroles:
    enum:
    - Admin
    - Manager
    - Member
    type: string
    x-enum-varnames:
    - UserrolesAdmin
    - UserrolesManager
    - UserrolesMember
  github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorBody:
    properties:
      body:
        example: error message
        type: string
    type: object
  github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse:
    properties:
      errors:
        $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorBody'
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API for managing teams, projects, tasks, and users
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Challenge FS Senior API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return access and refresh tokens
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_interfaces.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_interfaces.LoginResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "409":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
      summary: User login
      tags:
      - Auth
  /auth/logout:
    delete:
      consumes:
      - application/json
      description: Logout user and revoke all refresh tokens
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_interfaces.LogoutResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: User logout
      tags:
      - Auth
  /auth/refresh-token:
    post:
      consumes:
      - application/json
      description: Get a new access token using a refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_interfaces.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_interfaces.RefreshTokenResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "401":
          description: Invalid or expired refresh token
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
      summary: Refresh access token
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"